add tests for custom network builds
